Moved the code to the right file
I had the controller code in the models/poems file by accident.
